requirements.txt
attrs==22.1.0
autopep8==1.7.0
distro==1.7.0
func-timeout==4.3.5
fuzzy-match==0.0.1
fuzzysearch==0.7.3
jsonmerge==1.8.0
jsonschema==4.16.0
longest-increasing-subsequence==0.1.7
numpy==1.23.3
opencv-contrib-python==4.6.0.66
packaging==21.3
pandas==1.4.4
pdf2image==1.16.0
Pillow==9.2.0
pycodestyle==2.9.1
pyparsing==3.0.9
PyPDF2==2.10.6
pyrsistent==0.18.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
tabula-py==2.5.1
toml==0.10.2
